backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart

flask
flask_cors

python-socketio
python-jose
passlib[bcrypt]
uuid

requests
aiohttp
peewee
bcrypt

litellm
apscheduler
google-generativeai

langchain
langchain-community
chromadb
sentence_transformers
pypdf
docx2txt
unstructured
markdown
pypandoc
pandas
openpyxl
pyxlsb
xlrd

faster-whisper

PyJWT
pyjwt[crypto]

black
shutils
authlib
typing
fastapi_mail
itsdangerous
sentencepiece
transformers
